fix tests

+++ modules/70-dates/10-timestamp/index.php
@@ -7,6 +7,6 @@ const SECONDS_IN_YEAR = 60 * 60 * 24 * 365;
 function getYear($timestamp)
 {
     // BEGIN
-    return 1970 + floor($timestamp / SECONDS_IN_YEAR);
+    return 1970 + (int) floor($timestamp / SECONDS_IN_YEAR);
     // END
 }
